requirements.txt
feedparser>=6.0
requests>=2.28
beautifulsoup4>=4.11
lxml>=4.9
jinja2>=3.1
Flask>=3.0
python-dotenv>=1.0
google-genai>=1.0
APScheduler>=3.10
